requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pdfplumber>=0.7.0
pandas>=1.3.0
python-dotenv>=0.19.0
openai>=1.0.0
